fix shipping weight tests
